pyproject.toml
[project]
name = "database-prj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "matplotlib>=3.10.6",
    "mysql-connector-python>=9.4.0",
    "pandas>=2.3.2",
]
